requirements.txt
overpass
pyproj
pyyaml
shapely
